[Feature] Timeout cmd "Hello from Wemod Launcher"
This is the cmd that pops up that says "Hello from Wemod Launcher",
it would be nice if it could auto closes after 10 seconds, if nothing is done.
I assume that this cmd is added to confirm that wemod is working.
For that reason i assume that it would be fine if it auto closes.
It would be very convenient for the start to be automatic
and adding that timeout would do that. 
To make that happen i changed the "pause" command to the "choice" command.